gyp_skia: Fix module search path.
gyp_skia should prepend the local gyp source path to the python
module search path, instead of appending it. Otherwise, if a 'gyp'
module is locally installed on the machine, it will be read instead
of the current version in the Skia tree.

This breaks the Android build when the installed 'gyp' module doesn't
support the "make-android" generator.
Review URL: https://codereview.appspot.com/6344096

git-svn-id: https://skia.googlecode.com/svn/trunk@4832 2bbb7eff-a529-9590-31e7-b0007b416f81