Fixes for nacl_debugger's use of picture.
Now that CreateFromStream is used, check for NULL rather than
looking at width and height (prevents a crash on an invalid SKP).

Unref the picture when done with it.

Remove the member variable for picture, since it is only accessed
inside the one function.

R=borenet@google.com

Review URL: https://codereview.chromium.org/18055014

git-svn-id: https://skia.googlecode.com/svn/trunk@9831 2bbb7eff-a529-9590-31e7-b0007b416f81