Separate showing and hiding the debugger settings and inspector.
Internally the settings and inspector in the debugger are separate, but the
inspector button toggles both of them. This is unfortunate, because both of
these have large minimum vertical sizes. This means that those with vertical
resolutions less than ~1024 cannot see the the details tab.

This change separates the settings and inspector toggles into two so that
the interface is usable on screens of smaller size.

R=robertphillips@google.com

Review URL: https://codereview.chromium.org/18279002

git-svn-id: https://skia.googlecode.com/svn/trunk@9835 2bbb7eff-a529-9590-31e7-b0007b416f81