Add expectations files covering new image files.
New image files are in Google Storage, so add expectations for them.
The ICO file is expected to fail, although there is a patch to fix
that out for review (https://chromiumcodereview.appspot.com/18655007/)

Unreviewed.

Since we are currently ignoring failures in skimage, this should at
most produce a warning, but I expect it to remove the current warning
state.

The razr expectation is copied from nexus-7, as razr is currently
not working.

Review URL: https://codereview.chromium.org/18271012

git-svn-id: https://skia.googlecode.com/svn/trunk@9915 2bbb7eff-a529-9590-31e7-b0007b416f81