Return canonical logfont size to 64.
This is to fix crbug/277301. With skia:r9057 the canonical logfont size
was set to 2048 in an attempt to get better font metrics. The font size
is now being handled in a cleaner way, so this motivation no longer
exists. However, when the logfont size is large GDI will substitute a
vector font for a bitmap font. This results in incorrect results for
bitmap fonts as they are no longer properly detected by the typeface. 


git-svn-id: https://skia.googlecode.com/svn/trunk@11145 2bbb7eff-a529-9590-31e7-b0007b416f81