Always use GrOvalRenderer instead of GrDefaultPathRenderer for oval and rrect
BUG=https://codereview.chromium.org/271663002/

When drawing ovals with OpenGL device, skia uses GrOvalRenderer if multisampling
is disabled (fSampleCnt = 0) and use edge effect for antialiasing. If
multisampling is enabled, skia uses the default GrDefaultPathRenderer instead.

However, GrOvalRenderer is much faster (at least 300%) for drawing ovals and
circles even ignoring the multisampling and using edge effect for antialiasing.
For round rectangles it is also faster but almost negligibly.

This patch forces ovals and round rects to be drawn with GrOvalRenderer
regardless of the multisampling value.

Two new benchmarks are included: aaovals and aarrects.

Credits to Henry Song <henry.song@samsung.com> for spotting this.