Create reference/create a scratch texture from a bitmap only it does not have
A bitmap can be backed by a valid texture.  During when converting
a bitmap to a GrEffect from a bitmap shader, we check whether there is
 a valid texture.  Only create/reference a new texture if bitmap does not
have one.

If there is a texture, we must check whether target driver support npot and
tiling.  If not, we need to create a pot texture, paint orig texture to
pot texture