Do not optimize draw image if target driver does not support npot
If the target driver does not support npot image, we must create a
pot image, paint the original image to pot image.  In this case,
optimization of image draw does not have much sense because
we have to first draw the pot image. In this case, we disable image
drawing optimization.  Fortunately, most of modern driver (including
es3 drivers) support npot image tiling