Batch draw rectangle bitmap
There are three draw bitmap paths
1. draw bitmap via drawRect
2. draw bitmap via drawBitmapXXXX
3. draw bitmap via drawSprite

We can batch draw bitmap if two consequitive draws are the same bitmap.
What we need to do is set local rectangle of the bitmap.  The local matrix
of the bitmap is computed from the shader matrix and view matrix.  In case
of drawBitmap and drawSprite, there is no shader, so the shader matrix is
identity matrix.