[Skia] Batch draw rectangle bitmap.
There are three draw bitmap paths in Skia,
1. draw bitmap via drawRect
2. draw bitmap via drawBitmapXXXX
3. draw bitmap via drawSprite

Draw bitmap call can be batched if two consequitive draws are of same
bitmap. For this, local rectangle of the bitmap should be set.
The local matrix of the bitmap is computed from the shader matrix
and view matrix. In case of drawBitmap and drawSprite, there is no
shader, so the shader matrix is identity matrix.

Signed-off-by: suyambu.rm <suyambu.rm@samsung.com>