[Skia] Use approx equal to compare matrix for GrCoordTransform.
We batch drawBitmap, drawRect and drawOval/RRect by changing
GrProcessor's transformation matrix. SkMatrix has 9 floating points.
Concatinating SkMatrix and reverse it may result in loss of precision
during computation.

With exact match, we might get false negative when comparing two
GrCoordTransform matrices. We change that to use approxEqual of
SkMatrix. ApproxEqual() in SkMatrix is slower. But the extra time
in it should be small compare to the whole
drawRect/drawBitmap/drawOval/drawRRect.

Signed-off-by: Suyambulingam R M <suyambu.rm@samsung.com>