Create a scratch texture from a bitmap only if it does not have
A bitmap can be backed by a valid texture.  During converting a bitmap to
a GrFragmentProcessor, we check whether there is valid texture.  Only
create/reference a new texture if bitmap does not have one.

If the texture is of npot size, we also check whether driver supports NPOT
texture. We need to create a POT texture if driver does not support