[Skia] Create/reference a scratch texture from a bitmap only if it is not backed by a valid texture.
A bitmap can be backed by a valid texture. When converting
a bitmap to a GrFragmentProcessor from a bitmap shader, check whether
there is a valid texture associated with it. If not create/reference
a new texture. If there is a texture, we must check whether target
driver support npot and tiling. If not, we need to create a pot
texture and paint original texture to pot texture.

Signed-off-by: suyambu.rm <suyambu.rm@samsung.com>