Remove discardable memory based on size.  smaller size cache is removed first.
Reasoning behind this patch is that decoding smaller image is faster than
decoding larget image.  Thus, it is preferred to remove small image cache from
discardable memory pool instead of larger one.

However, there is a catch.  It is possible that larger image caches will
never be removed because of naiive technique of removing small cache first.

The remedy is add a timestamp (in milliseconds) when an image is added to
discardable memory pool.  during dumpDownTo() operation, we not only remove
those smaller image to meet the limit, but also we looks at those big images,
if these images are too old (3 seconds - this heuristic), we remove them as
well,

However, this patch can be controversial because the oldness is heuristic, we
might remove those images that in reality are still being used.  In this case,
we will get slower performance because we have re-decode those big images.