[Skia] Batch OvalRender for Bitmap Shader.
Add local coordinates in GrOvalRenderer. Local coordinates
describes the local coordinates for a bitmap in the shader.
Adding local coordinates enable us to batch drawCircle, drawEllipse
and drawRRect even if there are different view matrix in the canvas
and different local matrix in the shader.

Currently stroke is not batched in this case because of
changes in shader. (To be fixed)

Signed-off-by: Suyambulingam R M <suyambu.rm@samsung.com>